feat: Implementado Restrição de Operações para que Usuário só acesse seus Próprios Produtos

+++ backend/src/controllers/productController.js
@@ -4,8 +4,9 @@ const prisma = new PrismaClient();
 //CREATE
 export const createProduct = async (req, res) => {
     const { name, description, stock, price, category } = req.body;
+    const userId = req.user.id; //ID do Usuário logado.
     try {
-        const newProduct = await prisma.product.create({ data: { name, description, stock, price, category } });
+        const newProduct = await prisma.product.create({ data: { name, description, stock, price, category, userId } });
         res.json(newProduct);
     } catch (error) {
         console.error('Erro ao cadastrar produto.', error);
@@ -15,8 +16,9 @@ export const createProduct = async (req, res) => {
 
 //READ
 export const getAllProducts = async (req, res) => {
+    const userId = req.user.id;
     try {
-        const products = await prisma.product.findMany();
+        const products = await prisma.product.findMany({ where: { userId: userId } });
         res.json(products);
     } catch (error) {
         console.error('Erro ao buscar produtos.', error);
@@ -27,8 +29,12 @@ export const getAllProducts = async (req, res) => {
 //READ por ID
 export const getProductById = async (req, res) => {
     const { id } = req.params;
+    const userId = req.user.id;
     try {
-        const product = await prisma.product.findUnique({ where: { id: Number(id) } });
+        const product = await prisma.product.findFirst({ where: { id: Number(id), userId: userId } });
+        if (!product) {
+            return res.status(404).json({ error: 'Produto não encontrado.' });
+        }
         res.json(product);
     } catch (error) {
         console.error('Erro ao buscar produto.', error);
@@ -40,7 +46,12 @@ export const getProductById = async (req, res) => {
 export const updateProduct = async (req, res) => {
     const { id } = req.params;
     const { name, description, stock, price, category } = req.body;
+    const userId = req.user.id;
     try {
+        const product = await prisma.product.findFirst({ where: { id: Number(id), userId: userId } });
+        if (!product) {
+            return res.status(404).json({ error: 'Produto não encontrado ou não pertence a você.' });
+        }
         const updatedProduct = await prisma.product.update({ where: { id: Number(id) }, data: { name, description, stock, price, category } });
         res.json(updatedProduct);
     } catch (error) {
@@ -52,7 +63,12 @@ export const updateProduct = async (req, res) => {
 //DELETE
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
+    const userId = req.user.id;
     try {
+        const product = await prisma.product.findFirst({ where: { id: Number(id), userId: userId } });
+        if (!product) {
+            return res.status(404).json({ error: 'Produto não encontrado ou não pertece a você.' });
+        }
         await prisma.product.delete({ where: { id: Number(id) } });
         res.json({ message: 'Produto deletado!' });
     } catch (error) {
